// src/app/modules/cpassapi/model/settore.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Ente } from './ente';
import { TipoSettore } from './tipoSettore';
import { Utente } from './utente';


export interface Settore { 
    id?: string;
    codice?: string;
    descrizione?: string;
    indirizzo?: string;
    numCivico?: string;
    email?: string;
    contatto?: string;
    localita?: string;
    provincia?: string;
    cap?: string;
    telefono?: string;
    ente?: Ente;
    tipoSettore?: TipoSettore;
    utenteSettoreDefault?: boolean;
    rup?: Utente;
    idPadre?: string;
    livello?: number;
    listSettore?: Array<Settore>;
}