// src/app/modules/cpassapi/model/ente.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Ente { 
    id?: string;
    denominazione?: string;
    codiceFiscale?: string;
    codiceIpaAmministrazione?: string;
    dipartimento?: string;
    ufficio?: string;
    regione?: string;
    provincia?: string;
    indirizzo?: string;
    telefono?: string;
    email?: string;
    emailPEC?: string;
    dataCreazione?: Date;
    utenteCreazione?: string;
    dataModifica?: Date;
    utenteModifica?: string;
    dataCancellazione?: Date;
    utenteCancellazione?: string;
    optlock?: string;
}